Create LICENSE
Because I have other GPL 3.0 code in this branch that is currently unchanged (credit to https://github.com/JoshuaMKW/GeckoLoader for some file loading functionality in promptutils.py), after reading over the licensing I am adding a license to this program due to it using some code from there, as it seems that it is required to also license under GPL 3.0. I do not intend to sell this work or use it for commercial purposes, but I would rather be safe than sorry.